+++ src/components/grid-column/grid-sort-arrow.tsx
@@ -6,7 +6,7 @@ interface GridSortArrowProps {
 }
 
 export const GridSortArrow: FunctionalComponent<GridSortArrowProps> = ({ state }) => (
-    <svg style={{ marginLeft: 'auto' }} xmlns="http://www.w3.org/2000/svg" width="24" height="20" viewBox="0 0 24 20">
+    <svg aria-hidden="true" style={{ marginLeft: 'auto' }} xmlns="http://www.w3.org/2000/svg" width="24" height="20" viewBox="0 0 24 20">
         <style type="text/css">{`
             .arrow-group {
                 transform-origin: 50% 51%;